+++ packages/api/roles/all.js
@@ -42,6 +42,7 @@ export const all = (role, actions) => {
   role.allow(actions.updateOrderDelivery, () => false);
   role.allow(actions.updateOrderPayment, () => false);
   role.allow(actions.markOrderConfirmed, () => false);
+  role.allow(actions.markOrderRejected, () => false);
   role.allow(actions.markOrderPaid, () => false);
   role.allow(actions.markOrderDelivered, () => false);
   role.allow(actions.viewLogs, isInLoginMutationResponse);