+++ packages/core-orders/src/db/OrderStatus.ts
@@ -3,4 +3,5 @@ export enum OrderStatus {
   PENDING = 'PENDING',
   CONFIRMED = 'CONFIRMED',
   FULLFILLED = 'FULLFILLED',
+  REJECTED = 'REJECTED',
 }