+++ packages/platform/src/setup/setupTemplates.ts
@@ -1,4 +1,5 @@
 import { MessagingDirector } from 'meteor/unchained:core-messaging';
+import { resolveOrderRejectionTemplate } from '../templates/resolveOrderRejectionTemplate';
 import { resolveAccountActionTemplate } from '../templates/resolveAccountActionTemplate';
 import { resolveForwardDeliveryTemplate } from '../templates/resolveForwardDeliveryTemplate';
 import { resolveOrderConfirmationTemplate } from '../templates/resolveOrderConfirmationTemplate';
@@ -9,6 +10,7 @@ export const MessageTypes = {
   ACCOUNT_ACTION: 'ACCOUNT_ACTION',
   DELIVERY: 'DELIVERY',
   ORDER_CONFIRMATION: 'ORDER_CONFIRMATION',
+  ORDER_REJECTION: 'ORDER_REJECTION',
   QUOTATION_STATUS: 'QUOTATION_STATUS',
   ENROLLMENT_STATUS: 'ENROLLMENT_STATUS',
 };
@@ -17,6 +19,7 @@ export const setupTemplates = () => {
   MessagingDirector.registerTemplate(MessageTypes.ACCOUNT_ACTION, resolveAccountActionTemplate);
   MessagingDirector.registerTemplate(MessageTypes.DELIVERY, resolveForwardDeliveryTemplate);
   MessagingDirector.registerTemplate(MessageTypes.ORDER_CONFIRMATION, resolveOrderConfirmationTemplate);
+  MessagingDirector.registerTemplate(MessageTypes.ORDER_CONFIRMATION, resolveOrderRejectionTemplate);
   MessagingDirector.registerTemplate(MessageTypes.QUOTATION_STATUS, resolveQuotationStatusTemplate);
   MessagingDirector.registerTemplate(MessageTypes.ENROLLMENT_STATUS, resolveEnrollmentStatusTemplate);
 };